Add support for conditional graphs
New function generateConditionalGraphs.
Uses the same querying structure as dynamic graphs, but inserts all matches as
a list of values in each target.

Adds support to create dashboards with graphs which are based on a query (e.g.
existance of a type of metric which means some specific application runs on
the server) but graph values which are not directly implicit on the query.

Fixes #9