Remove last 90 seconds from aggregated graphs
On graphs which calculate aggregations over a lot of metrics we see sometimes a
big drop in the last part of the graph. This is due to the lack of some of the
metrics' last points (for not all rrd's are updated syncronously).

By removing the last 90 seconds from these graphs the drop is no longer plotted
which prevents false alarms, while loosing the last 3 data points seem to be an
acceptable drawback.