Add dashboard list dictionary for real paths
Allow arbitrary organization of dashboards nested in sub-directories.

Entries on dashboard-index.json should include a slash to indicate a sub-directory, and
a dot for normal files. Both separators will be converted in a tree structure.

The ```Graphitus.Tree``` object now has a pathInfo array which keeps indexed the dashboard
path in a normalized way (with dots) along with its real path information as defined on
```dashboards-index.json```.

Thus, dashboards filed under:
```
dashboards/core-team/dht.overview.json
```
will be included as:
```
{"id": "core-team/dht.overview"}
```
in the ```dashboard-index.json``` file.

The tree will be constructed as:
```
core-team ---> dht ---> overview
```