core-team: Replace size+records for logfiles with disk usage
The total size and number of records reported by the node is completely
inaccurate so it doesn't make much sense to show it (is just an upper
bound, as when records are removed from disk the internal counters are
never updated).

Also add the plain disk usage (not as percentage) to the overview and
aggregates, and a new dashboard for both disk and disk-pcnt, to match the
dashboards in the dht about memory usage.