Move new dashboards to subdirectories
The core-team and adserving-team dashboards are moved to subdirectories
in the dashboards directory for easier maintenance.

Also, the "ad-serving" dashboard was renamed to "adserving-team"
dashboard so it matches the GitHub team name too, in case at some point
we want to make some link, it should be easier this way.

Finally, the dashboards are reordered in the index. The new dashboards
(per team) are placed at the top of the tree and sorted alphabetically).
The old dashboards are left with the order untouched after the new
dashboards.