Remove timeShift
We used to timeShift by 45s to stop people panicing when they saw the graphs
plummet through delayed data in graphite. It seems data is less delayed now
but also keepLastValue now works so we use that instead.

A better solutiuon would be for graphitus to not draw leave blanks for
unknown values but that is a bit harder.