core: Show memory usage per server in overviews
In the team and dht overview, each individual node was being graphed in
the memory percentage graph, which doesn't provide very useful
information, as it doesn't give a proper overview on how full the whole
servers are, which is what's more important.

This commit changes this to show the sum of all the nodes on each
server.

Fixes #136.